added return list of deleteds

+++ src/use-cases/logins-has-permissions/delete/remove-all-logins-by-permission-delete.usecase.ts
@@ -1,3 +1,4 @@
+import { loginsHasPermissions } from "@prisma/client";
 import { RecordNotFound } from "src/errors/prisma/record-not-found.error";
 import { InterfaceLoginHasPermissionRepository } from "src/repository/interfaces/interface-login-has-permission.repository";
 import { InterfaceLoginRepository } from "src/repository/interfaces/interface-login.repository";
@@ -13,14 +14,21 @@ class RemoveAllLoginsByPermissionDeleteUseCase {
     private permissionRepository: InterfacePermissionRepository    
   ) {}  
   
-  async execute({permission}: RemoveAllLoginByPermissionDeleteUseCaseRequest): Promise<void> {
+  async execute({permission}: RemoveAllLoginByPermissionDeleteUseCaseRequest) {
     const permissionExists = await this.permissionRepository.findByName(permission);
 
     if (!permissionExists) {
       throw new RecordNotFound(permission);
-    }
+    }    
 
-    return await this.loginHasPermissionRepository.removeAllLoginsByPermission(permissionExists.uuid);
+    const loginsByPermission = await this.loginHasPermissionRepository.getAllLoginsByPermission(permissionExists.uuid);
+
+    if (loginsByPermission.length == 0) {
+      throw new RecordNotFound("logins");
+    }
+    
+    await this.loginHasPermissionRepository.removeAllLoginsByPermission(permissionExists.uuid);
+    return loginsByPermission;
   }
 }
