+++ src/use-cases/logins-has-permissions/delete/remove-all-permissions-by-login-delete.usecase.ts
@@ -1,3 +1,4 @@
+import { loginsHasPermissions } from "@prisma/client";
 import { RecordNotFound } from "src/errors/prisma/record-not-found.error";
 import { InterfaceLoginHasPermissionRepository } from "src/repository/interfaces/interface-login-has-permission.repository";
 import { InterfaceLoginRepository } from "src/repository/interfaces/interface-login.repository";
@@ -13,14 +14,21 @@ class RemoveAllPermissionsByLoginDeleteUseCase {
     private loginRepository: InterfaceLoginRepository,    
   ) {}
 
-  async execute({email}: RemoveAllPermissionByLoginDeleteUseCaseRequest): Promise<void> {
+  async execute({email}: RemoveAllPermissionByLoginDeleteUseCaseRequest): Promise<loginsHasPermissions[]> {
     const userExists = await this.loginRepository.findByEmail(email);
     
     if (!userExists) {
       throw new RecordNotFound("email");
     }
     
-    return await this.loginHasPermissionRepository.removeAllPermissionsByLogin(userExists.uuid);
+    const permissionsByLogin = await this.loginHasPermissionRepository.getAllPermissionsByLogin(userExists.uuid);
+    
+    if(permissionsByLogin.length == 0) {
+      throw new RecordNotFound("permissions");
+    }
+
+    await this.loginHasPermissionRepository.removeAllPermissionsByLogin(userExists.uuid);
+    return permissionsByLogin;
   }
 }
